==> Notebooks/1_etl_limpieza_movies_subset.ipynb <==
# 1.- ETL
## 1.1 PASO 1. Proceso de limpieza de datos segun los requerimientos mínimos según Henry
1. Eliminar los campos que no se utilizaran: 'video','imdb_id','adult', 'original_title', 'poster_path', 'homepage'.
2. Dar formato de AAAA-mm-dd de haber fecha en el campo 'release_date' y crear el campo 'release_year'.
3. Eliminar los valores nulos del campo 'release_date'.
4. Eliminar los valores nulos de los campos 'revenue', 'budget' deben ser rellenados por el número 0.
5. Crear el campo 'return' con los campos 'revenue' y 'budget', dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0.
6. Desanidar los campos: 'belongs_to_collection', 'genres', 'production_companies', 'production_countries' y 'spoken_languages'.

import pandas as pd
import numpy as np
import os

data_movies = pd.read_csv('D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Datasets\movies_dataset.csv')
data_movies.rename(columns={'id':'movie_id'}, inplace=True) #Cambiamos el id a movie_id para poder usarlo como primary key
data_movies.info()


1. Eliminar los campos que no se utilizaran: 'video','imdb_id','adult', 'original_title', 'poster_path', 'homepage'.

columnas_excluidas_henry = ['video','imdb_id','adult','original_title','poster_path', 'homepage'] # Columnas que no serán utilizadas segun Henry
columnas_excluidas_criterio_personal = ['tagline', 'status']    #Columnas que no seran utilizadas segun criterio personal    
columnas_desanidar=['belongs_to_collection','genres','production_companies','production_countries','spoken_languages'] #Columnas que requieren ser desanidadas

data_movies_subset = data_movies.drop(list(columnas_excluidas_henry+columnas_excluidas_criterio_personal+columnas_desanidar), axis=1)
data_movies_subset.info()

#data_movies_subset['movie_id']=pd.to_numeric(data_movies_subset['movie_id'])

data_movies_subset.iloc[19730]['movie_id'] 


# identificamos las filas que contienen datos tipo fecha en la columna de nuestro primary key: pelicula_id
def check_type(x):
     try: 
        return type(eval(x)) 
     except Exception as e:
        return type(x)

data_movies_subset.loc[data_movies_subset['movie_id'].map(check_type)!=int]

# removemos las filas que contienen fecha en la columna de primary key
data_movies_subset.drop(
    data_movies_subset.loc[data_movies_subset['movie_id'].map(check_type)!=int].index,
    inplace=True)
print("Numero de filas que no contienen datos tipo fecha en la columna ['movie_id']:",
    data_movies_subset.loc[data_movies_subset['movie_id'].map(check_type)!=int].shape[0]
)

#data_movies_subset['movie_id']=pd.to_numeric(data_movies_subset['movie_id'])
#data_movies_subset.head()

2. Dar formato de AAAA-mm-dd de haber fecha en el campo 'release_date' y crear el campo 'release_year'.

#Convertimos el tipo de columna
columns_numeric=['movie_id','budget','revenue','popularity','vote_average','vote_count','runtime']
columns_date=['release_date']

data_movies_subset[columns_numeric]=data_movies_subset[columns_numeric].apply(pd.to_numeric) # convertimos a tipo numerico
data_movies_subset[columns_date]=data_movies_subset[columns_date].apply(pd.to_datetime) # convertimos a tipo fecha 

data_movies_subset_limpia=data_movies_subset.convert_dtypes() # volvemos a evaluar el df
data_movies_subset_limpia.info()
data_movies_subset_limpia.head(3)

3. Eliminar los valores nulos del campo 'release_date'.

# instruccion 1: Los valores nulos del campo release date deben eliminarse
data_movies_subset_limpia['release_date'].isna().any()
print("Cantidad de valores nulos en ['release_date']:",
    data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].shape[0])
data_movies_subset_limpia['release_date'].isna()
rows_release_date_na = data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].index
data_movies_subset_limpia.drop(rows_release_date_na, inplace=True)
print("Cantidad de valores nulos en ['release_date'] despues del drop:",
    data_movies_subset_limpia.loc[data_movies_subset_limpia['release_date'].isna()].shape[0])
#Crear el 'release_year'
data_movies_subset_limpia['release_year'] = data_movies_subset_limpia['release_date'].dt.year

4. Eliminar los valores nulos de los campos 'revenue', 'budget' deben ser rellenados por el número 0.

# Reemplazar valores nulo de las columnas 'revenue' y 'budget' con "0"
data_movies_subset['revenue'] = data_movies_subset['revenue'].fillna(0)
data_movies_subset['budget'] = data_movies_subset['budget'].fillna(0)
# Cantidad de valores nulos en ['budget','revenue']
data_movies_subset[['budget','revenue']].isna().sum()

5. Crear el campo 'return' con los campos 'revenue' y 'budget', dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0.

#Creación de columna 'return' = 'revenue' / 'budget'
data_movies_subset_limpia['return'] = data_movies_subset_limpia['revenue']/data_movies_subset_limpia['budget']

#Reemplazo de valores nulos e indefinidos con '0'
data_movies_subset_limpia['return'] = data_movies_subset_limpia['return'].replace([np.inf, -np.inf], 0).fillna(0)
data_movies_subset_limpia.info()

print("Hay valores nulos en cualquiera de ['revenue','budget','return']:", any(data_movies_subset_limpia[['revenue','budget','return']].isna().any()))


# Removemos duplicados
data_movies_subset_limpia.drop_duplicates(inplace=True) # no funciona; deben existir filas que no son exactamente iguales.
print("Numero de peliculas unicas:",data_movies_subset_limpia['movie_id'].nunique(),"\n")
print("Numero de registros (filas):",data_movies_subset_limpia.shape[0],"\n")
print("Numero de peliculas con informacion inconsiste:",data_movies_subset_limpia.shape[0]-data_movies_subset_limpia['movie_id'].nunique(),"\n")

# para una misma pelicula existen valores distintos.
data_movies_subset_inconsistent=data_movies_subset_limpia.loc[data_movies_subset_limpia.duplicated(subset=['movie_id'], keep=False)]
print("Numero de filas con informacion duplicada en algunas columnas:",
    data_movies_subset_inconsistent.shape[0])
print("Numero de peliculas unicas con informacion duplicadas en algunas columnas:",
    data_movies_subset_inconsistent['movie_id'].nunique())
data_movies_subset_inconsistent.sort_values(['movie_id']).head(4)

# Solucionamos tomando la primera entrada para cada pelicula, y descartamos el resto.
data_movies_subset_limpia_sindups=data_movies_subset_limpia.loc[~data_movies_subset_limpia.duplicated(subset=['movie_id'], keep='first')]
print("Numero de peliculas unicas:",data_movies_subset_limpia_sindups['movie_id'].nunique(),"\n")
print("Numero de registros (filas):",data_movies_subset_limpia_sindups.shape[0],"\n")
print("Numero de peliculas con informacion inconsiste:",
      data_movies_subset_limpia_sindups.shape[0]-data_movies_subset_limpia_sindups['movie_id'].nunique(),"\n")

Guardamos el dataset limpio

# Nuestro dataset final
print("Universo de peliculas con datos unicos:",data_movies_subset_limpia_sindups['movie_id'].nunique(),"\n")
data_movies_subset_limpia_sindups.to_csv(os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Pipeline","data_movies_subset_limpia.csv"))
data_movies_subset_limpia_sindups.info()
data_movies_subset_limpia_sindups.head(2)
