==> Notebooks/2_etl_desanidado_movies.ipynb <==
## Desanidar y normalizar valores de movies
En esta parte desanidamos y normalizamos las columnas de movies_dataset.csv

import pandas as pd
from pandas import json_normalize
import ast
import json
import os
import zipfile
from ast import literal_eval

data_movies = pd.read_csv('D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Datasets\movies_dataset.csv')
data_movies.rename(columns={'id':'movie_id'}, inplace=True) #Cambiamos el id a movie_id para poder usarlo como primary key
data_movies.info()

a. Columna: belongs_to_collection

# Segun el diccionario entregado nos concentramos en: belongs_to_collection
data_movies_collection = pd.DataFrame(data_movies[['belongs_to_collection','movie_id']]).dropna()
data_movies_collection.head()

# tercero se normaliza la columna
data_movies_collection=data_movies_collection.set_index('movie_id')

# Reemplazar NaN en la columna 'belongs_to_collection' por un diccionario predeterminado
data_movies_collection['belongs_to_collection'] = data_movies_collection['belongs_to_collection'].apply(lambda x: {'id': 0, 'name': 'Desconocido', 'poster_path': 'N/A', 'backdrop_path': 'N/A'} if pd.isna(x) else x)

# Convierte de string a diccionario
data_movies_collection['belongs_to_collection'] = data_movies_collection['belongs_to_collection'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)

# Desanida la columna 'belongs_to_collection'
data_movies_collection_unnested = json_normalize(data_movies_collection['belongs_to_collection']).set_index(data_movies_collection.index)
#data_movies_collection_unnested = pd.json_normalize(data_movies_collection['belongs_to_collection']).set_index(data_movies_collection.index)

data_movies_collection_unnested.head()

# Cambiamos el nombre de las columnas
data_movies_collection_unnested.rename(columns = {'id':'franquicia_id','name':'franquicia'}, inplace = True)
data_movies_collection_unnested.reset_index(inplace=True) # mueve el indice a una columna
data_movies_collection_unnested.info()

b. Columna: production_companies

# nos concentramos en: production_companies
data_movies_companies=pd.DataFrame(data_movies[['production_companies','movie_id']]).dropna()
data_movies_companies.head()

data_movies_companies['production_companies']=data_movies_companies['production_companies'].apply(literal_eval)
data_movies_companies.head()

# segundo, expandimos la columna para que queden diccionarios anidados (deja de ser una lista de diccionarios)
data_movies_companies_exploded = data_movies_companies.explode('production_companies', ignore_index=True)
data_movies_companies_exploded.head(2)

# tercero se normaliza la columna
data_movies_companies_exploded=data_movies_companies_exploded.set_index('movie_id')
data_movies_companies_exploded_unnested = pd.json_normalize(data_movies_companies_exploded['production_companies']).set_index(data_movies_companies_exploded.index)
data_movies_companies_exploded_unnested.head(2)

# se finaliza el df con el cambio de nombre de columna.
data_movies_companies_exploded_unnested.rename(columns = {'id':'productora_id','name':'productora'}, inplace = True)
data_movies_companies_exploded_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_movies_companies_exploded_unnested.head(2)

c. Columna: genres

# nos concentramos en:  genres 
data_movies_genres=pd.DataFrame(data_movies[['genres','movie_id']]).dropna()
data_movies_genres.head(2)

# replicamos los pasos para el caso de  lista de diccionario
data_movies_genres['genres']=data_movies_genres['genres'].apply(literal_eval)
data_movies_genres_exploded = data_movies_genres.explode('genres', ignore_index=True)
data_movies_genres_exploded=data_movies_genres_exploded.set_index('movie_id')
data_movies_genres_exploded_unnested = pd.json_normalize(data_movies_genres_exploded['genres']).set_index(data_movies_genres_exploded.index)
data_movies_genres_exploded_unnested.rename(columns = {'id':'genres_id','name':'genres'},inplace = True)
data_movies_genres_exploded_unnested.reset_index(inplace=True) # mueve le indice a una columna
data_movies_genres_exploded_unnested.head(2)


data_movies_countries = pd.DataFrame(data_movies[['production_countries', 'movie_id']]).dropna()
data_movies_countries['production_countries'] = data_movies_countries['production_countries'].apply(literal_eval)
data_movies_countries_exploded = data_movies_countries.explode('production_countries')
data_movies_countries_exploded_unnested = pd.json_normalize(data_movies_countries_exploded['production_countries']).set_index(data_movies_countries_exploded.movie_id)
data_movies_countries_exploded_unnested.reset_index(inplace=True)
data_movies_countries_exploded_unnested.rename(columns={'iso_3166_1':'pais_isocode','name' : 'pais_name'},inplace=True)
data_movies_countries_exploded_unnested.head(2)

data_movies_languages = pd.DataFrame(data_movies[['spoken_languages', 'movie_id']]).dropna()
data_movies_languages['spoken_languages']= data_movies_languages['spoken_languages'].apply(literal_eval)
data_movies_languages_exploded = data_movies_languages.explode('spoken_languages')
data_movies_languages_exploded_unnested = pd.json_normalize(data_movies_languages_exploded['spoken_languages']).set_index(data_movies_languages_exploded.movie_id)
data_movies_languages_exploded_unnested.reset_index(inplace=True)
data_movies_languages_exploded_unnested.rename(
    columns={'iso_639_1':'spoken_languages_isocode','name' : 'spoken_languages_name'},
    inplace=True
    )
data_movies_languages_exploded_unnested.head(2)

# unir los dataframes
data_movies_normalized = pd.merge(
    pd.merge(pd.merge(pd.merge(data_movies_collection_unnested,
    data_movies_companies_exploded_unnested,
    on=['movie_id']),
    data_movies_genres_exploded_unnested,
    on=['movie_id']),
    data_movies_countries_exploded_unnested,
    on=['movie_id']),
    data_movies_languages_exploded_unnested,
    on=['movie_id']
    )
data_movies_normalized.head(2)

data_movies_final=data_movies_normalized[
    data_movies_normalized.columns[~data_movies_normalized.columns.str.contains("_path")] # quitamos las NO contienen "_path"
    ]
data_movies_final.to_csv(os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Pipeline","data_movies_normalizada.csv"))
data_movies_final.info()
