==> Notebooks/4.etl_union_datafinal.ipynb <==
# Consolidamos los df para alimentar las consultas
En esta etapa consolidamos los dataframes generados con anterioridad. 

import pandas as pd
import os
data_movies_subset = pd.read_csv(
    os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Pipeline","data_movies_subset_limpia.csv"),index_col=0).convert_dtypes() # output paso 01
data_movies_normalizada = pd.read_csv(
    os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Pipeline","data_movies_normalizada.csv"),index_col=0).convert_dtypes() # output paso 02
data_credits_normalizada = pd.read_csv(
    os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Pipeline","credits_normalizada.csv"),index_col=0).convert_dtypes() # output paso 03

data_movies_subset.info(verbose=True)
data_movies_subset.isna().sum()

data_movies_normalizada.info(verbose=True)
data_movies_normalizada.isna().sum()

data_credits_normalizada.info(verbose=True)
data_credits_normalizada.isna().sum()
#data_credits_normalizada.head(10)

# Descripcion de la disponibilidad de la informacion 

# evaluamos cual df contiene informacion para el mayor numero de casos.
movies_subset_peliculas=data_movies_subset['movie_id'].nunique()
movies_normalizada_peliculas=data_movies_normalizada['movie_id'].nunique()
credits_normalizada_peliculas=data_credits_normalizada['movie_id'].nunique()
print("Universo de pelicula unicas, segun df de referencia:",movies_subset_peliculas)
print("Universo de pelicula unicas, segun movies_normalizada:",movies_normalizada_peliculas)
print("Universo de pelicula unicas, segun credits_normalizada:",credits_normalizada_peliculas)
print("Porcentaje de valores nulos en columnas de movies_normalizada del total de registros (df referencia):",
    round((1-(movies_normalizada_peliculas/movies_subset_peliculas))*100,2),"%")
print("Porcentaje de valores nulos en columnas de credits_normalizada del total de registros (df referencia):",
    round((1-(credits_normalizada_peliculas/movies_subset_peliculas))*100,2),"%")

## Implicaciones de disponibilidad de datos para las consultas
Las funciones que mas se afectarian por esta baja disponibilidad de datos seria las siguientes consultas:

- productoras existosas  
- pais de produccion  
- director  
- sistema de recomendacion (modelo ml)  
Con base en lo anterior, desarrollamos un estrategia de imputacion de valores nulos.

## Imputacion de valores nulos

# exportamos el df final
data_mvp_final_normalizada = pd.merge(
    pd.merge(data_movies_subset, # tomamos como principal referencia el df con el mayor numero de registros
    data_movies_normalizada,
    how='left',
    on=['movie_id']),
    data_credits_normalizada, # tomamos como ultima referencia el df con el menor numero de registros
    how='left',
    on=['movie_id'])
#data_mvp_final_normalizada.info()
data_mvp_final_normalizada.isna().sum()

# Evaluamos la cantidad de peliculas sin idioma
print("Numero de peliculas sin idioma:",data_mvp_final_normalizada.loc[
    data_mvp_final_normalizada['original_language'].isna()==True]['movie_id'].nunique())

# guardamos la posicion de las filas para comparar el resultado de la imputacion
language_null_index=data_mvp_final_normalizada.loc[
    data_mvp_final_normalizada['original_language'].isna()==True].index

# visualizamos
data_mvp_final_normalizada.loc[
    data_mvp_final_normalizada['original_language'].isna()==True].head(3)

# Imputamos en idioma con base en el titulo de la pelicula
from langdetect import detect, DetectorFactory
DetectorFactory.seed = 0
print("Ejemplo de identificacion de idioma para 'Lettre d'une inconnue':",detect("Lettre d'une inconnue"))

language_fill = data_mvp_final_normalizada.loc[
    data_mvp_final_normalizada['original_language'].isna()==True, 'title'].apply(detect) # aplicamos la funcion de identificacion de idiomas a los casos nulos
data_mvp_final_normalizada.loc[
    data_mvp_final_normalizada['original_language'].isnull() & (data_mvp_final_normalizada['original_language'].isna()==True), # condiciones
    'original_language'] = language_fill # imputamos los valores al base de datos

# visualizamos los cambios realizados
data_mvp_final_normalizada[['movie_id','title','original_language']].loc[language_null_index].drop_duplicates()

# imputamos valores nulos en runtime usando la mediana, puesto que es mas resistente a la influencia de outliers
imp_median_runtime=data_mvp_final_normalizada[['runtime']].drop_duplicates().dropna().median()
data_mvp_final_normalizada[['runtime']] = data_mvp_final_normalizada[['runtime']].fillna(imp_median_runtime)

# asumimos que los valores nulos en franquicia corresponden a peliculas que no pertnenecen a ninguna franquicia.
# Con base en lo anterior, imputamos el valor 'Sin franquicia' para los casos nulos
data_mvp_final_normalizada['franquicia'].fillna('Sin franquicia', inplace = True)

columns_desconocido=['productora','genres','pais_name','spoken_languages_name','director']
data_mvp_final_normalizada.loc[:,columns_desconocido]=data_mvp_final_normalizada.loc[:,columns_desconocido].fillna('desconocido')

# Nuestros resultados despues de la primera fase de imputacion
#data_mvp_final_normalizada.info()
data_mvp_final_normalizada.isna().sum()

## 4.2 Generacion de dfs para alimentar las consultas
Separamos el df final en dos subconjuntos:  
1. df para alimentar las funciones
2. df para alimentar el modelo ml y el sistema de recomendacion.

### 4.2.1. Subconjunto de datos para alimentar las funciones


# primer subconjunto: df para alimentar las funciones
columns_funciones=[
    'movie_id', # todas
    'title', # varias
    'original_language', # def peliculas_idioma( Idioma: str )
    'runtime', # def peliculas_duracion( Pelicula: str )
    'release_date', # def peliculas_duracion( Pelicula: str )
    'franquicia', # def franquicia( Franquicia: str )
    'budget', # def franquicia( Franquicia: str )
    'revenue', # def franquicia( Franquicia: str )
    'pais_name', # def peliculas_pais( Pais: str )
    'productora', # productoras_exitosas( Productora: str ),
    'director', # get_director( nombre_director )
    'return', # get_director( nombre_director )
    'release_year' # get_director( nombre_director )
    ]

data_mvp_funciones=data_mvp_final_normalizada.loc[:,columns_funciones].drop_duplicates()
print("Universo de peliculas unicas en el df que alimenta las funciones:",
      data_mvp_funciones['movie_id'].nunique())

data_mvp_funciones.to_csv(os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Output","data_mvp_final_funciones.csv"))
data_mvp_funciones.info()
data_mvp_funciones.isna().sum()

### 4.2.3. Subconjunto de datos para alimentar la consulta de director

# primero, calculamos el retorno por director
data_mvp_funciones_exitodir=data_mvp_funciones[['director','revenue','budget']].loc[
    (data_mvp_funciones['revenue']!=0) & (data_mvp_funciones['budget']!=0) # filtramos unicamente aquellas filas para las que hay informacion
].drop_duplicates()
print("Numero de directores con informacion disponible", data_mvp_funciones_exitodir['director'].nunique())
data_exitodir_revenue=data_mvp_funciones_exitodir[['director','revenue']].groupby(['director']).sum()
data_exitodir_budget=data_mvp_funciones_exitodir[['director','budget']].groupby(['director']).sum()
data_mvp_funciones_exitodir=pd.merge(
    data_exitodir_revenue,
    data_exitodir_budget,
    on=['director'])
data_mvp_funciones_exitodir['return']=data_mvp_funciones_exitodir['revenue']/data_mvp_funciones_exitodir['budget']
data_mvp_funciones_exitodir.rename(
    columns={'return': 'director_return',
             'revenue':'director_revenue',
             'budget':'director_budget'},
    inplace=True)

# exportamos
data_mvp_funciones_exitodir.reset_index(inplace=True)
data_mvp_funciones_exitodir.info()

# segundo subconjunto: df para alimentar las funciones
columns_exitodir=[
    'movie_id', # todas
    'title', # varias
    'release_date', # def peliculas_duracion( Pelicula: str )
    'release_year',
    'budget', # def franquicia( Franquicia: str )
    'revenue', # def franquicia( Franquicia: str )
    'return', # get_director( nombre_director )
    'director', # get_director( nombre_director )
    ]

data_mvp_funciones_exitodir_2join=data_mvp_final_normalizada.loc[:,columns_exitodir].drop_duplicates()
data_mvp_exitodir=data_mvp_funciones_exitodir.merge(
    data_mvp_funciones_exitodir_2join,
    on=['director']
)

data_mvp_exitodir.to_csv(os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Output","data_mvp_final_funciones_exitodir.csv"))
data_mvp_exitodir.info()
data_mvp_exitodir.isna().sum()

### 4.2.2. Subconjunto de datos para alimentar el modelo ML

# segundo subconjunto: df para alimentar el modelo de ML
columns_ml=[
    'movie_id',
    'title',
    'popularity',
    'vote_average',
    'vote_count', # Coneteo de votos
    'release_year',
    'genres',
    'runtime',
    #'pais_name',
    #'franquicia',
    'original_language',
    'prtgnst_gender',
    'prtgnst_nivel',
    #'director',
    #'executive_producer',
    'overview'
    ]

# aprovechamos las columnas ya limpias en data_mvp_funciones que tambien estan en columns_ml:
columns_funciones_2join=list(set(columns_ml) & set(columns_funciones))
print(columns_funciones_2join)
# tambien incluimos las columnas en columns_ml que no han sido limpiadas:
columns_ml_2join=list(['movie_id']+list(set(columns_ml) - set(columns_funciones)))
print(columns_ml_2join)

# subconjunto de datos para modelo ml
data_mvp_ml=pd.merge(
      data_mvp_funciones.loc[:,columns_funciones_2join].drop_duplicates(),
      data_mvp_final_normalizada.loc[:,columns_ml_2join].drop_duplicates(),
      on='movie_id',
      how='left')
print("Universo de peliculas unicas en el df que alimenta el modelo ml:",
      data_mvp_funciones['movie_id'].nunique())

# identificamos las columnas con nulos
data_mvp_ml.info()
data_mvp_ml.isna().sum()

#### 4.2.2.2. Imputacion de valores 'desconocido'  
Nos concentramos en las columnas restantes que contienen nulos.


columns_fillna_num=[
    'prtgnst_nivel']
columns_fillna_cat=[
    'prtgnst_gender']
data_mvp_ml[columns_fillna_num+columns_fillna_cat].isna().sum()

El website the IMDB señala que su fuente de datos no es unicamente estudios de produccion y realizadores de cine, sino que "[...] la mayor parte de nuestra información procede de profesionales y visitantes" (IMDB).

Por tal motivo, imputamos el valor desconocido en las variables categoricas, y el valor 99 en las numericas con el fin de facilitar el filtro de informacion mas adelante.

# algunos casos de peliculas que pertenecen al grupo "peliculas con baja circulacion"
index_fillna_cat=data_mvp_ml.loc[data_mvp_ml['prtgnst_gender'].isna()==True].index # guardamos para comparar
data_mvp_ml[['movie_id','title']+columns_fillna_cat].loc[data_mvp_ml['prtgnst_gender'].isna()==True].head(3)

data_mvp_ml.tail()

# inputamos valores nulos con las columnas seleccionadas con el valor 'baja_circulacion';
# cambiamos prtgnst_gender a string para facilitar a imputacion de valores
data_mvp_ml['prtgnst_gender'] = data_mvp_ml['prtgnst_gender'].astype('object')
data_mvp_ml['prtgnst_gender']=data_mvp_ml['prtgnst_gender'].fillna('desconocido')

data_mvp_ml[['movie_id','title']+columns_fillna_cat].loc[index_fillna_cat].head(3)

data_mvp_ml[columns_fillna_num].drop_duplicates()

# inputamos valores nulos con las columnas seleccionadas con el valor 'baja_circulacion';
# cambiamos prtgnst_nivel a string para facilitar a imputacion de valores
data_mvp_ml['prtgnst_nivel']=data_mvp_ml['prtgnst_nivel'].astype('float')
data_mvp_ml['prtgnst_nivel']=data_mvp_ml['prtgnst_nivel'].fillna(99)
data_mvp_ml[columns_fillna_num].drop_duplicates()

# Para facilitar nuestro modelo de ML,
# filtramos nuestros datos para concentrarnos unicamnete en el genero del protagonista
# filtramos 'prtgnst_nivel' == 1; es decir, excluimos actores secundarios
data_mvp_ml=data_mvp_ml.loc[data_mvp_ml['prtgnst_nivel'].isin([1])]

# El resultado despues de completar todas las fases de imputacion: exportamos el df para alimentar el modelo ML
data_mvp_ml.to_csv(os.path.join("D:\Bootcamp\Henry\CURSO\Proyecto Individual 1\Movies\Output","data_mvp_final_ml.csv"))
data_mvp_ml.info()
data_mvp_ml.isna().sum()
